requirements.txt
djangorestframework
psycopg2
django
pymongo
drf-yasg
pillow
django-cors-headers
djangorestframework-simplejwt
django-celery-beat
django-filter
reportlab
requests
openpyxl
textstat
celery
djongo
python-decouple
gunicorn
